chapter_003/src/test/java/ru.job4j/list/ConvertListTest.java
package ru.job4j.list;

import org.junit.Test;

import java.util.ArrayList;
import java.util.Arrays;
import java.util.List;

import static org.hamcrest.core.Is.is;
import static org.junit.Assert.assertThat;

public class ConvertListTest {
    @Test
    public void when7ElementsThen9() {
        ConvertList list = new ConvertList();
        int[][] result = list.toArray(
                Arrays.asList(1, 2, 3, 4, 5, 6, 7),
                3
        );
        int[][] expect = {
                {1, 2, 3},
                {4, 5, 6},
                {7, 0, 0}
        };
        assertThat(result, is(expect));
    }

    @Test
    public void whenConvertListOfMassiveToList() {
        ConvertList list = new ConvertList();
        List<Integer[]> input = new ArrayList<Integer[]>();
        input.add(new Integer[]{1, 2});
        input.add(new Integer[]{3, 4, 5, 6});
        List<Integer> result = list.convert(input);
        List<Integer> expect = Arrays.asList(
                1, 2, 3, 4, 5, 6
        );
        assertThat(result, is(expect));
    }
}